Update test files

+++ tests/expected/equations.jl
@@ -1,7 +1,7 @@
 
 if abspath(PROGRAM_FILE) == @__FILE__
-    x::Int64 = 0
-    y::Int64 = 0
+    x = 0
+    y = 0
     @assert(!(x > 2))
     @assert(y < 10)
     @assert((x + 2 * y) == 0)
