+++ tests/expected/unary_op.jl
@@ -1,5 +1,5 @@
 if abspath(PROGRAM_FILE) == @__FILE__
-    a::Int64 = 2
+    a = 2
     @assert(~a == -3)
     -1
     +1
